backend/pyproject.toml
[project]
name = "praxia-backend"
version = "0.1.0"
description = "PraxIA - AI Agent Studio Backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Redis
    "redis[hiredis]>=5.2.0",
    # Task queue
    "arq>=0.26.1",
    # OpenAI
    "openai>=1.57.0",
    "tiktoken>=0.8.0",
    # Auth (JWT verification done manually, pyjwt for production)
    "pyjwt[crypto]>=2.9.0",
    # HTTP client
    "httpx>=0.28.0",
    # Logging
    "structlog>=24.4.0",
    # YAML
    "pyyaml>=6.0.2",
    # CORS
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
